+++ R/grid_terrain.r
@@ -32,8 +32,6 @@
 #' Interpolate ground points and create a rasterized digital terrain model. The interpolation
 #' can be done using 3 methods: \code{"knnidw"}, \code{"delaunay"} or \code{"kriging"} (see
 #' details). The algorithm uses the points classified as "ground" to compute the interpolation.
-#' The function forces the original lowest ground point of each pixel (if it exists) to be
-#' chosen instead of the interpolated values.
 #'
 #' \describe{
 #' \item{\code{knnidw}}{Interpolation is done using a k-nearest neighbour (KNN) approach with
@@ -52,8 +50,10 @@
 #' @param res numeric resolution.
 #' @param method character can be \code{"knnidw"}, \code{"delaunay"} or \code{"kriging"} (see details)
 #' @param k numeric. number of k-nearest neighbours when the selected method is either \code{"knnidw"} or \code{"kriging"}
-#' @param model a variogram model computed with \link[gstat:vgm]{vgm} when the selected method is \code{"kriging"}.
-#' If null it performs an ordinary or weighted least squares prediction.
+#' @param model a variogram model computed with \link[gstat:vgm]{vgm} when the selected method
+#' is \code{"kriging"}. If null it performs an ordinary or weighted least squares prediction.
+#' @param keep_lowest logical. The function forces the original lowest ground point of each
+#' pixel (if it exists) to be chosen instead of the interpolated values.
 #' @return A \code{lasmetrics} data.table.
 #' @export
 #' @examples
@@ -81,7 +81,7 @@
 #' \link[akima:interp]{interp}
 #' \link[lidR:lasnormalize]{lasnormalize}
 #' \link[raster:raster]{RasterLayer}
-grid_terrain = function(.las, res = 1, method, k = 10L, model = gstat::vgm(.59, "Sph", 874))
+grid_terrain = function(.las, res = 1, method, k = 10L, model = gstat::vgm(.59, "Sph", 874), keep_lowest = FALSE)
 {
   . <- X <- Y <- Z <- NULL
 
@@ -101,9 +101,12 @@ grid_terrain = function(.las, res = 1, method, k = 10L, model = gstat::vgm(.59, 
 
   grid[, Z := round(Zg, 3)]
 
-  # force ground point to be dominant
-  grid = rbind(grid, grid_metrics(lasfilterground(.las), list(Z = min(Z)), res))
-  grid = grid[, list(Z = min(Z)), by = .(X,Y)]
+  # force lowest ground point to be dominant
+  if(keep_lowest)
+  {
+    grid = rbind(grid, grid_metrics(lasfilterground(.las), list(Z = min(Z)), res))
+    grid = grid[, list(Z = min(Z)), by = .(X,Y)]
+  }
 
   as.lasmetrics(grid, res)